Fix for bug 103799.  Saved the volume with double precision fidelity and feed that back to the control system.
Code Review: Rob

git-svn-id: svn://linux2/subversion/Crestron/Android_tst600/trunk/customFiles/src/external/crestron/CresStreamSvc@203865 634c09e3-b06c-4293-a491-0870fed93cb0

+++ src/com/crestron/txrxservice/AudioPlayback.java
@@ -90,7 +90,7 @@ public class AudioPlayback
 //        	                        mPlayer.flush();
         	                        if (initVolumePending)
         	                        {
-        	                        	setVolume(mStreamCtl.userSettings.getVolume());
+        	                        	setVolume((int)mStreamCtl.userSettings.getVolume());
         	                        	initVolumePending = false;
         	                        }
                             	}
